feat(monobot): Generate versions.json files
monobot takes as input a config file (currently monobot.json) and
generates output files. The end goal is generating a repo.json file.

This commit is a intermediate step towards generating a more complete
repo.json file. As of this commit, this file is actually not generated.
Instead, a versions.json file is generated, which is a part of the final
repo.json.

versions.json is interesting because it's completely machine generated
from the inspection of the tags (and possibly commits) of the remote
repository. It is kept as a separate file for future commits to parse it
before running the bot so that known versions are not re-scanned again.

The final repo.json will be assembled (at a later commit) from some
static configuration that is user-provided via the input config file
(monobot.json) and one or more machine-generated config files, among
which versions.json is the first supported one.

ConfigObjectMapperCustomizer: customized ObjectMapper that writes map
keys in order (as Monogres Bazel scripts require to keep order as of
now).

Minor refactors and code improvement in the affected paths.