feat(monobot): Incremental generation of versions.json
This commit introduces support to parse versions.json for existing
versions and avoiding fetching and extracting the metadata for versions
that were already present in versions.json.

Add short_commit additional property to each version.

Ensure correctness in keeping the order of the entries in version, when
doing a merge between existing versions and newly detected versions.

GitTag: moved logic to extract a list of tags from a given repo to a
static method.