feat(monobot): Revamp monobot.json config format, support metadata propagation
The main goal is to make monobot.json file format as easy as it can
gets, so that friction to contribute new components (mainly extensions)
to the repo is as small as possible.

Thanks to the use of some defaults (configuration by convention) we can
simplify the monobot.json configuration to the following simple JSON
config:

{
    "name": "noset",
    "url": "https://gitlab.com/ongresinc/extensions/noset"
}

For the above, type=EXTENSION and version=V1 are added as default
values.

An additional (and optional) component of monobot.json is the metadata
section. Added here and wrote code to propagate it to the resulting
repo.json, in case it's present.