feat(monobot): Add a parser for Postgres extension .control files
Control is a record with properly validated and understanding of the
default values of a Postgres extension control file. It is modeled after
the official documentation:

https://www.postgresql.org/docs/current/extend-extensions.html

It also includes the code for parsing the file from an array of bytes.

Control uses CsvStringArrayDeserializer to deserialize values that are
an array of strings to a String[].